Finish restructuring to create-react-app v2, next edit the UI
