+++ src/components/StepIndicator.js
