+++ src/views/Offices/Detail/index.js
@@ -1,0 +1,28 @@
+import { connect } from 'react-redux';
+
+import KanimDetail from '../components/KanimDetail';
+import { confirmOfficeQuota, confirmQuotaSync } from '../actions/kanim';
+import { registerQueue } from '../actions/queue';
+import { addNotification, viewNotifModifySync } from '../actions/notif';
+
+const mapStateToProps = ({ kanim, auth, queue, notif }) => ({
+  auth,
+  office: kanim.office,
+  confirmation: kanim.confirmation,
+  getOfficeQuotaAttempt: kanim.getOfficeQuotaAttempt,
+  confirmQuotaAttempt: kanim.confirmQuotaAttempt,
+  registerQueueAttempt: queue.registerQueueAttempt,
+  addNotificationAttempt: notif.addNotificationAttempt
+});
+const mapDispatchToProps = {
+  confirmOfficeQuota,
+  confirmQuotaSync,
+  registerQueue,
+  addNotification,
+  viewNotifModifySync
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(KanimDetail);
